# PharmacyOnline/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PharmacyOnline, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Profile/candidate/create": {
      "post": {
        "tags": [
          "Profile"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProfileModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/update": {
      "post": {
        "tags": [
          "Profile"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idProfileDetail",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProfileModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/submit": {
      "post": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCandidate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "idProfileDetail",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/cancel": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCandidate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "idProfileDetail",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/delete": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCandidate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "idProfileDetail",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/getprofile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCandidate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/candidate/gethistory": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCandidate",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/approved": {
      "post": {
        "tags": [
          "Profile"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/approvedCV"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/search": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/getlist/submitted": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/getlist/history": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isQualified",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/getdetail": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdProfile",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Profile/admin/download/fileCV": {
      "get": {
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idProfileDetail",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProfileModel": {
      "required": [
        "fullname",
        "idCandidate",
        "number"
      ],
      "type": "object",
      "properties": {
        "idCandidate": {
          "format": "int32",
          "type": "integer"
        },
        "fullname": {
          "minLength": 1,
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "number": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "thumbnail": {
          "format": "binary",
          "type": "string"
        },
        "fileCv": {
          "format": "binary",
          "type": "string"
        },
        "skills": {
          "type": "string"
        },
        "universityOrCollege": {
          "type": "string"
        },
        "major": {
          "type": "string"
        },
        "issuedDate": {
          "format": "date-time",
          "type": "string"
        },
        "expiryDate": {
          "format": "date-time",
          "type": "string"
        },
        "scientificAchievements": {
          "type": "string"
        },
        "workExperiences": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "age": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "gender": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "approvedCV": {
      "type": "object",
      "properties": {
        "idProfileDetail": {
          "type": "string"
        },
        "isQualified": {
          "format": "int32",
          "type": "integer"
        },
        "body": {
          "$ref": "#/definitions/emailModel"
        }
      },
      "additionalProperties": false
    },
...
    "emailModel": {
      "type": "object",
      "properties": {
        "to": {
          "format": "email",
          "type": "string"
        },
        "interviewAddress": {
          "type": "string"
        },
        "appointment": {
          "format": "date-time",
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}